Create shared and static libequistore directly
Previously we would create these as libequistore_core.so, and rename the
file to libequistore.so. This is much harder to do on Windows, where
the .dll.lib file contains an hardcoded `equistore_core.dll`, and modifying the
.dll.lib file is non-trivial.

Since we actually don't need to build the rlib for equistore-core, we can
set lib.name to `equistore` and everything works just fine!