Remove IndexError backtrace for last iteration step over Labels
Using C10_THROW_ERROR is relatively slow, since it need to compute a
full backtrace. When iterating over TorchLabels, Python will call __getitem__
with increasing integers until an IndexError is raised. In our case, the
creation of the IndexError is taking a lot more time than the iteration
itself. The fix here is to directly throw a manually constructed IndexError.